backend/requirements.txt
fastapi
uvicorn
python-dotenv
ffmpeg-python
openai
ctransformers
whisper

